% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankroll_plot.R
\name{bankroll_plot}
\alias{bankroll_plot}
\title{Bankroll Simulation}
\usage{
bankroll_plot(
  bets = 256,
  win_rate = 0.55,
  bet_size = 100,
  sim_length = 1000,
  avg_odds = -110,
  odds_type = "us",
  current_bet = NULL,
  current_win = NULL
)
}
\arguments{
\item{bets}{The number of bets (256)}

\item{win_rate}{The average expected win rate of the bets (0-1)}

\item{bet_size}{The dollar amount of each bet. (100)}

\item{sim_length}{The number of simulations. (1,000)}

\item{avg_odds}{The average odds of the bets (-110)}

\item{odds_type}{Type of odds for the output ("us", "dec", "frac")}

\item{current_bet}{Optional input - Your current total number of tracked bets. (125)}

\item{current_win}{Optional input - Your current total amount won/loss from your tracked bets. (950)}
}
\value{
Plot showing the bankroll results of the simulated bets. The function also outputs the percentage of positive and negative final bankroll over the course of the simulation.
}
\description{
This function simulates and visualizes a bettors bankroll over a number of bets using their edge.
}
\examples{
bankroll_plot(bets = 256, win_rate = 0.55, bet_size = 100, sim_length = 1000, avg_odds = -110, odds_type = "us")
bankroll_plot(sim_length = 500, avg_odds = -110, win_rate = 0.5455)
bankroll_plot(sim_length = 250, avg_odds = -115, win_rate = 0.5255, current_bet = 100, current_win = -500)
bankroll_plot(sim_length = 300, avg_odds = -109, win_rate = 0.57, current_bet = 175, current_win = 5000)
bankroll_plot()
}
