% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_calc.R
\name{bet_calc}
\alias{bet_calc}
\title{Calculate Bet Payout}
\usage{
bet_calc(risk, odds, type = "us")
}
\arguments{
\item{risk}{Amount risked for your bet(s) (110, 105)}

\item{odds}{Odds for the bet (-132, -115)}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Oddss
}}
}
\value{
payout Payout of a bet
}
\description{
This function calculates the payout for a given bet.
}
\examples{
bet_calc(risk = 100, odds = -132, type = "us")
bet_calc(risk = 500, odds = 2.3, type = "dec")
bet_calc(risk = 25, odds = 20/1, type = "frac")

}
