% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_prob.R
\name{bet_prob}
\alias{bet_prob}
\title{Bet Probability}
\usage{
bet_prob(pred_spread, spread, sport = "NBA")
}
\arguments{
\item{pred_spread}{Predicted spread for the team you want to bet on}

\item{spread}{Spread for the team that you want to bet on (-3.5, -7, 4, 2.5)}

\item{sport}{Sport/League of the teams being bet on ("NBA", "NCAAB", "NFL", "NCAAF")}
}
\value{
probs
}
\description{
This function returns the win, loss, and push probabilities of a given bet based on the current line and your predicted line.
}
\examples{
bet_prob(-9, -3.5, sport = "NFL")
bet_prob(-7, -3, sport = "NBA")
bet_prob(21, 10.5, sport = "NCAAF")
bet_prob(-3, 5, sport = "NCAAB")
}
