% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet_prob_plot.R
\name{bet_prob_plot}
\alias{bet_prob_plot}
\title{Bet Probability Plot}
\usage{
bet_prob_plot(pred_spread, spread, sport = "NBA")
}
\arguments{
\item{pred_spread}{Predicted spread for the team you want to bet on}

\item{spread}{Spread for the team that you want to bet on (-3.5, -7, 4, 2.5)}

\item{sport}{Sport/League of the teams being bet on. Possible values are:
\itemize{
\item \code{NBA}, National Basketball Association
\item \code{NCAAB}, College Basketball
\item \code{NFL}, National Football League
\item \code{NCAAF}, College Football
}}
}
\value{
probs Plot of the simulation as well as the percentage of simulations that were positive and negative.
}
\description{
This function plots the win, loss, and push probabilities of a given bet based on the current line and your predicted line.
}
\examples{
bet_prob_plot(-9, -3.5, sport = "NFL")
bet_prob_plot(-7, -3, sport = "NBA")
bet_prob_plot(21, 10.5, sport = "NCAAF")
bet_prob_plot(-3, 5, sport = "NCAAB")

}
\references{
Stern, Hal. "The Probability of Winning a Football Game as a function of the Pointspread." The American Statistician 45, no. 3 (1991): 179-83. Accessed July 18, 2020. doi:10.2307/2684286. \url{https://statistics.stanford.edu/sites/g/files/sbiybj6031/f/COV\%20NSF\%2059.pdf}

Stern, Hal. "On the Probability of Winning a Football Game." The American Statistician 45, no. 3 (1991): 179-83. Accessed July 18, 2020. doi:10.2307/2684286. \url{https://www-jstor-org.turing.library.northwestern.edu/stable/2684286}

Winston, Wayne L. "From Point Ratings to Probabilities." In Mathletics: How Gamblers, Managers, and Sports Enthusiasts Use Mathematics in Baseball, Basketball, and Football, 290-97. PRINCETON; OXFORD: Princeton University Press, 2009. Accessed July 18, 2020. doi:10.2307/j.ctt7sj9q.48.
}
