% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_even.R
\name{break_even}
\alias{break_even}
\title{Break Even Probability}
\usage{
break_even(risk, rtrn, legs = 1)
}
\arguments{
\item{risk}{Amount risked on the bet}

\item{rtrn}{Amount returned on the bet. (Risk + the amount won in the bet)}

\item{legs}{The number of legs needed to win the bet. (straight bets = 1, parlay >= 2)}
}
\value{
prob
}
\description{
This function returns the break even win percentage needed based on the amount risked and the amount returned on the bet.
It also factors in parlay bets with the number of legs needed to convert to payout.
Gives you the percent chance you need to break even on the bet
}
\examples{
break_even(110, 210, 1)
break_even(50, 750, 4)
}
