% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hfa.R
\name{calc_hfa}
\alias{calc_hfa}
\title{Calculate a team's Home Field Advantage}
\usage{
calc_hfa(win_p_neutral, win_p_leag_hm)
}
\arguments{
\item{win_p_neutral}{percentage of time team A beats team B on neutral field.}

\item{win_p_leag_hm}{league average win percentage for home teams.}
}
\value{
numeric scalar between 0 and 1
}
\description{
Calculate the percentage of time team A beats team B at home based of the
percentage of time team A wins on a neutral field and the league average win
percentage at home.
}
\examples{
calc_hfa(.142, .61)

}
\references{
See post by Ganchrow at the SBR forum for details \href{https://www.sportsbookreview.com/forum/handicapper-think-tank/122312-home-field-advantage.html}{Home field advantage}
}
