% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clv_calc.R
\name{clv_calc}
\alias{clv_calc}
\title{Closing Line Value}
\usage{
clv_calc(bet_odds, closing_odds, type = "us")
}
\arguments{
\item{bet_odds}{Odds for your bet(s) (-110, -105)}

\item{closing_odds}{Closing Odds for the same bet (-132, -115)}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}
}
\value{
clv Closing Line Value of your bet
}
\description{
This function calculates the Closing Line Value (CLV) of your bets.
}
\examples{
clv_calc(
  bet_odds = -110,
  closing_odds = -132,
  type = "us"
)
clv_calc(
  bet_odds = 2.5,
  closing_odds = 2.3,
  type = "dec"
)
clv_calc(
  bet_odds = 50/1,
  closing_odds = 20/1,
  type = "frac"
)

}
