% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_calc.R
\name{edge_calc}
\alias{edge_calc}
\title{Calculate Edge}
\usage{
edge_calc(win_prob, odds, type = "us")
}
\arguments{
\item{win_prob}{Your predicted probability for winning the bet (0-1)}

\item{odds}{The odds for the given bet}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}
}
\value{
edge The edge that you have for the given bet.
}
\description{
This function will calculate the edge of a given bet based on your predicted win probability of the bet and the bet's odds.
}
\examples{
edge_calc(0.75, -175, type = "us")
edge_calc(0.75, 1.2855, type = "dec")
edge_calc(0.75, 3/7, type = "frac")
edge_calc(
  win_prob = c(0.6, 0.7, 0.52, 0.6),
  odds = c(-110, -150, 140, 150),
  type = "us"
)

}
