% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_scores.R
\name{expected_scores}
\alias{expected_scores}
\title{Expected Scores}
\usage{
expected_scores(home_line, total)
}
\arguments{
\item{home_line}{Line for the home team (-6)}

\item{total}{Total for the game (212)}
}
\value{
dataframe with expected home score and away score for the game
}
\description{
This function calculates the expected scores for a given game.
}
\examples{
expected_scores(home_line = -6, total = 212)
expected_scores(home_line = -2.5, total = 51)
expected_scores(home_line = 10.5, total = 231)

}
