% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_value.R
\name{expected_value}
\alias{expected_value}
\title{Expected Value}
\usage{
expected_value(payout, risk, win_prob)
}
\arguments{
\item{payout}{The dollar amount won if the bet pays out (500)}

\item{risk}{The dollar amount risked (550)}

\item{win_prob}{The probability that the bet wins (0.54)}
}
\value{
Expected Return on Investment
}
\description{
This function calculates the expected value for a given bet.
}
\examples{
expected_value(100, 110, 0.55)
expected_value(175, 100, 0.35)
}
