% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_odds.R
\name{fair_odds}
\alias{fair_odds}
\title{True Expected Probability}
\usage{
fair_odds(line = -110, odds = c(-110, -110), type = "us")
}
\arguments{
\item{line}{Line for the specific bet (-115)}

\item{odds}{Vector with all lines for a wager (-115, -105)}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}
}
\value{
True Expected probability of bet
}
\description{
This function calculates the fair odds for a bet.
}
\examples{
fair_odds(1.8, c(1.8, 2.1), type = "dec")
fair_odds(-115, c(-115, -105))
fair_odds(258, c(285, -122, 258))
fair_odds(line = -100, odds = c(-100, -120))

}
