% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_nba_names.R
\name{fix_nba_names}
\alias{fix_nba_names}
\title{Fix NBA Team Names}
\usage{
fix_nba_names(x)
}
\arguments{
\item{x}{String - NBA Team name (Baltimore Ravens)}
}
\value{
Abbreviated NBA team name.
}
\description{
This function changes NBA team names to their abbreviated team name.
}
\examples{
fix_nba_names("atlanta")
fix_nba_names("HAWKS")


}
