% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_ncaaf_names.R
\name{fix_ncaaf_names}
\alias{fix_ncaaf_names}
\title{Fix NCAAF Team Names}
\usage{
fix_ncaaf_names(x)
}
\arguments{
\item{x}{String - NCAAF Team name (Utah Utes)}
}
\value{
Abbreviated NCAAF team name.
}
\description{
This function changes NCAAF team names to their abbreviated team name.
}
\examples{
fix_ncaaf_names("arizona")
fix_ncaaf_names("CARDINALS")


}
