% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_nfl_names.R
\name{fix_nfl_names}
\alias{fix_nfl_names}
\title{Fix NFL Team Names}
\usage{
fix_nfl_names(x)
}
\arguments{
\item{x}{String - NFL Team name (Baltimore Ravens)}
}
\value{
Abbreviated NFL team name.
}
\description{
This function changes NFL team names to their abbreviated team name.
}
\examples{
fix_nfl_names("arizona")
fix_nfl_names("CARDINALS")


}
