% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lines.R
\name{get_lines}
\alias{get_lines}
\title{Scrape SportsBook Review's Historical Betting Lines}
\usage{
get_lines(
  sport = "NFL",
  bet_type = "spread",
  period = "full",
  start_date = "20191222"
)
}
\arguments{
\item{sport}{Sport to scrape. Possible values are:
\itemize{
\item \code{NFL}, National Football League
\item \code{NBA}, National Basketball Association
\item \code{NHL}, National Hockey League
\item \code{MLB}, Major League Baseball
\item \code{NCAAF}, College Football
\item \code{NCAAB}, College Basketball
}}

\item{bet_type}{Type of bets. Possible values are:
\itemize{
\item \code{spread},
\item \code{total},
\item \code{moneyline},
}}

\item{period}{Length of bet. Possible values are:
\itemize{
\item \code{full}, Full game
\item \verb{1H}, 1st Half
\item \verb{2H}, 2nd Half
\item \verb{1Q}, 1st Quarter
\item \verb{2Q}, 2nd Quarter
\item \verb{3Q}, 3rd Quarter
\item \verb{4Q}, 4th Quarter
}}

\item{start_date}{Date of sporting events}
}
\value{
dataframe containing betting lines for the given day
}
\description{
This function scrapes the sportsbookreview.com website to return betting lines
}
\examples{
get_lines(
  sport = "NFL",
  bet_type = "spread",
  period = "full",
  start_date = "20191222"
)
get_lines(
  sport = "NBA",
  bet_type = "total",
  period = "2Q",
  start_date = "20191221"
)
get_lines(
  sport = "NCAAF",
  bet_type = "moneyline",
  period = "1H",
  start_date = "20191019"
)

}
