% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hold_calc.R
\name{hold_calc}
\alias{hold_calc}
\title{Sportsbook Hold Percentage}
\usage{
hold_calc(...)
}
\arguments{
\item{...}{Vector of lines for a given bet (-115, -105)}
}
\value{
Hold percent
}
\description{
This function calculates the hold perrcentage that the sportsbook has for the given bet.
}
\examples{
hold_calc(-110, -110)
hold_calc(285, -122, 258)
}
