% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_odds.R
\name{implied_odds}
\alias{implied_odds}
\title{Probability of Winning Conversion to Odds (American Odds)}
\usage{
implied_odds(prob, type = "us")
}
\arguments{
\item{prob}{Probability of winning a bet}

\item{type}{Odds Type ("us", "dec", "frac") "us" == American Odds, "dec" == Decimal Odds, "frac" == Fractual Odds}
}
\value{
odds American Odds of that bet
}
\description{
This function provides the fair odds for a given win probability for a given bet.
}
\examples{
implied_odds(0.4)
implied_odds(0.5238095)
implied_odds(c(0.3, 0.2, 0.909, 0.7143), type = "dec")
implied_odds(c(0.3, 0.2, 0.95, 0.7), type = "frac")
}
