% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_prob.R
\name{implied_prob}
\alias{implied_prob}
\title{Implied Probability of Winning (American Odds)}
\usage{
implied_prob(odds, type = "us")
}
\arguments{
\item{odds}{American Odds of the bet}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}
}
\value{
Implied probability of winning or the "Break Even" percentage
}
\description{
This function provides the implied probability or "Break Even" win percentage for a given bet.
}
\examples{
implied_prob(-150)
implied_prob(360)
implied_prob(c(360, -500, -110, 140))
implied_prob(c(3.60, 5.00, 1.10, 1.40), type = "dec")
implied_prob(c(5/2, 1/2, 7/1, 20/1, 10/11), type = "frac")

}
