% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelly.R
\name{kelly}
\alias{kelly}
\title{Kelly Criterion Calculation}
\usage{
kelly(win_prob, odds, type = "dec", kelly_type = "full")
}
\arguments{
\item{win_prob}{Probability of winning a bet (0.55)}

\item{odds}{Odds for a bet (-110)}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}

\item{kelly_type}{Optional input altering the kelly formula. Possible values are:
\itemize{
\item \code{half}, Half Kelly
\item \code{quarter}, Quarter Kelly
\item \code{eighth}, Eighth Kelly
}}
}
\value{
Percentage of bankroll to risk on bet
}
\description{
This function calculates the Kelly Criterion percentage of your bankroll to bet in order to maximize returns.
}
\examples{
kelly(win_prob = 0.58, odds = -132, type = "us")
kelly(win_prob = 0.53, odds = -105, type = "us", kelly_type = "half")
kelly(win_prob = 0.545, odds = 2.1, type = "dec")
kelly(win_prob = 0.27, odds = 5.5, type = "dec", kelly_type = "quarter")
kelly(win_prob = 0.10, odds = 40/1, type = "frac")

}
\references{
\href{https://en.wikipedia.org/wiki/Kelly_criterion}{Kelly Criterion wikipedia} page
}
