% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelly_bet.R
\name{kelly_bet}
\alias{kelly_bet}
\title{Kelly Criterion Bet Size}
\usage{
kelly_bet(unit_size, win_prob, odds, type = "dec", kelly_type = "full")
}
\arguments{
\item{unit_size}{Unit size of your bankroll, typically 1\% of bankroll (100)}

\item{win_prob}{Probability of winning a bet (0.55)}

\item{odds}{Odds for a bet (-110)}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}

\item{kelly_type}{Optional input altering the Kelly formula. Possible values are:
\itemize{
\item \code{half}, Half Kelly
\item \code{quarter}, Quarter Kelly
\item \code{eighth}, Eighth Kelly
}}
}
\value{
Percentage of bankroll to risk on bet
}
\description{
This function calculates the Kelly Criterion and returns the dollar amount to bet in order to maximize returns.
}
\examples{
kelly_bet(
  unit_size = 100,
  win_prob = 0.58,
  odds = -132,
  type = "us"
)
kelly_bet(
  unit_size = 50,
  win_prob = 0.53,
  odds = -105,
  type = "us",
  kelly_type = "Half"
)
kelly_bet(
  unit_size = 500,
  win_prob = 0.545,
  odds = 2.1,
  type = "dec"
)
kelly_bet(
  unit_size = 300,
  win_prob = 0.27,
  odds = 5.5,
  type = "dec",
  kelly_type = "Quarter"
)
kelly_bet(
  unit_size = 250,
  win_prob = 0.10,
  odds = 40/1,
  type = "frac"
)

}
\references{
\url{https://en.wikipedia.org/wiki/Kelly_criterion}
}
