% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_round.R
\name{over_round}
\alias{over_round}
\title{Sportsbook's Over Round Percentage}
\usage{
over_round(...)
}
\arguments{
\item{...}{Lines for a given bet (-115, -105)}
}
\value{
Bet's Over Round percent
}
\description{
This function calculates the extra implied probability from each line of the bet. House Edge or Over Round.
}
\examples{
over_round(-110, -110)
over_round(-125, -125)
over_round(285, -122, 258)

}
