% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlay_calc.R
\name{parlay_calc}
\alias{parlay_calc}
\title{Parlay Calculation}
\usage{
parlay_calc(risk, odds, type = "us")
}
\arguments{
\item{risk}{Odds for your bet(s)}

\item{odds}{Vector of odds for each leg of the parlay (-132, -115, -110, -110)}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}
}
\value{
payout Parlay payout
}
\description{
This function calculates the payout of parlay bets.
}
\examples{
parlay_calc(risk = 75, odds = c(-110, -110, -110), type = "us")
parlay_calc(risk = 50, odds = c(-110, -110, -110, -110), type = "us")
parlay_calc(risk = 100, odds = c(-110, -150, -175, -325, -220), type = "us")

}
