% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_bet.R
\name{random_bet}
\alias{random_bet}
\title{Random Bet Result}
\usage{
random_bet(risk = 110, payout = 100, num_bets = 1, win_rate = 0.55)
}
\arguments{
\item{risk}{The dollar amount risked to place the bet}

\item{payout}{The dollar amount the bet pays out if it wins}

\item{num_bets}{Number of bets to simulate}

\item{win_rate}{The average expected win rate of the bets (0-1)}
}
\value{
The dollar amount won/lost from the bet
}
\description{
This function simulates a bet
}
\examples{
convert_odds(c(-110, -110))
convert_odds(odds = c(1.1, 2.1, 13, 6.5, 1.909), input = "dec", output = "us")
convert_odds(odds = c(1/10, 11/10, 12/1, 11/2, 10/11), input = "frac", output = "all")
}
