% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_implied_prob.R
\name{true_implied_prob}
\alias{true_implied_prob}
\title{True Implied Probability of Winning (American Odds)}
\usage{
true_implied_prob(odds, type = "us")
}
\arguments{
\item{odds}{American Odds of the bet}

\item{type}{Type of odds. Possible values are:
\itemize{
\item \code{us}, American Odds
\item \code{dec}, Decimal Odds
\item \code{frac}, Fractional Odds
}}
}
\value{
Implied probability of winning or the "Break Even" percentage
}
\description{
This function provides the true implied win probability for a bet.
}
\examples{
true_implied_prob(-150)
true_implied_prob(360)
true_implied_prob(c(360, -500, -110, 140))
true_implied_prob(c(3.60, 5.00, 1.10, 1.40), type = "dec")
true_implied_prob(c(5/2, 1/2, 7/1, 20/1, 10/11), type = "frac")

}
