% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_odds.R
\name{convert_odds}
\alias{convert_odds}
\title{Odds converter}
\usage{
convert_odds(odds, input = "us", output = "all")
}
\arguments{
\item{odds}{Vector of lines for a given bet (-115, -105)}

\item{input}{Vector of lines for a given bet ("us", "dec", "frac", "prob")}

\item{output}{Vector of lines for a given bet ("all", "us", "dec", "frac", "prob")}
}
\value{
odds
}
\description{
This function converts any odds or probability.
}
\examples{
convert_odds(c(-110, -110))
convert_odds(odds = c(1.1, 2.1, 13, 6.5, 1.909), input = "dec", output = "us")
convert_odds(odds = c(1/10, 11/10, 12/1, 11/2, 10/11), input = "frac", output = "all")
}
