% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_round.R
\name{over_round}
\alias{over_round}
\title{Sportsbook's Over Round Percentage}
\usage{
over_round(...)
}
\arguments{
\item{...}{Lines for a given bet (-115, -105)}
}
\value{
Hold percent
}
\description{
This function calculates the extra implied probability from each line of the bet. House Edge or Over Round.
}
\examples{
hold_calc(-110, -110)
hold_calc(-125, -125)
hold_calc(285, -122, 258)

}
